+++ CHANGES.rst
@@ -19,8 +19,10 @@ everywhere, in favor of the new :attr:`pyudev.Device.action`:
 
 - Deprecated :attr:`pyudev.Monitor.enable_receiving` in favor of
   :attr:`pyudev.Monitor.start`.
-- Added :attr:`pyudev.Monitor.started`, :attr:`pyudev.Monitor.start` can be
+- Added :attr:`pyudev.Monitor.started`, :meth:`pyudev.Monitor.start` can be
   called repeatedly now
+- Added :meth:`pyudev.Monitor.poll`, and deprecated
+  :meth:`pyudev.Monitor.receive_device`
 
 
 0.15 (Mar 1, 2012)